requirements.txt
google-cloud-pubsub
apache-beam[gcp]
matplotlib
pandas
google-cloud-bigquery
google-cloud-storage
